back-character-forge/src/main/java/com/jawbr/dnd5e/characterforge/controller/ClassController.java
package com.jawbr.dnd5e.characterforge.controller;

import com.jawbr.dnd5e.characterforge.dto.response.FindAllDTOResponse;
import com.jawbr.dnd5e.characterforge.dto.response.theClass.ClassDTO;
import com.jawbr.dnd5e.characterforge.service.ClassService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/classes")
public class ClassController {

    private final ClassService classService;

    public ClassController(ClassService classService) {
        this.classService = classService;
    }

    @GetMapping
    public FindAllDTOResponse findAllClasses() {
        return classService.findAllClasses();
    }

    @GetMapping("/{indexName}")
    public ClassDTO findClassByIndexName(@PathVariable String indexName) {
        return classService.findClassByIndexName(indexName);
    }

}
